# Speeding up compilation of the tests that rely on Catch2