--- a/tests/DGtalCatch.h
+++ b/tests/DGtalCatch.h
@@ -6,6 +6,5 @@
 **/
 
 #pragma once
-#define CATCH_CONFIG_MAIN
 #define CATCH_CONFIG_ENABLE_BENCHMARKING
 #include <catch2/catch.hpp>