# PR Description

While using Catch2 in another project, I realize that test compilation was really slow. I didn't take care of that in DGtal since DGtal itself is a huge code.

A first hint is in the Catch2 documentation (https://github.com/catchorg/Catch2/blob/devel/docs/slow-compiles.md) proposing to move the `#define CATCH_CONFIG_MAIN` directive in a dedicated source file (`DGtalCatch.cpp` in this PR) and making all test file linking against it.

Doing so speedup compilation time of a test file after each modification but one `DGtalCatch.o` is still compiled for each test. This [answer](https://stackoverflow.com/a/1388969) suggests to compile this file as a dedicated library (`DGtalCatch` in this PR).

A little benchmark using following commands:
```bash
cmake .. -DBUILD_TESTING=TRUE
time make testPointVector testLinearizer testNumberTraits testPointFunctorHolder testUnorderedSetByBlock testHyperRectDomain testPointHashFunctions testProfile testMeaningfulScaleAnalysis testMultiStatistics testLambdaMST3DBy2D testParametricCurveDigitization testDigitalSurfaceRegularization testShroudsRegularization testIntegralInvariantShortcuts testSphericalHoughNormalVectorEstimator testDigitalConvexity testLpMetric testBoundedRationalPolytope testBoundedLatticePolytope testCellGeometry testConvexHull2DReverse testConvexHull2DThickness testSetFunctions testSimpleRandomAccessRangeFromPoint testFunctorHolder testContainerTraits testSurfaceMesh testTriangulatedSurface testMeshVoxelization testProjection testLemniscate2D testAstroid2D testPolygonalSurface testShapeMoveCenter testHalfPlane testObjectBoostGraphInterface testKhalimskySpaceND testParDirCollapse testHalfEdgeDataStructure testIndexedDigitalSurface testVoxelComplex testCubicalComplex testNeighborhoodConfigurations testConstImageFunctorHolder testArrayImageAdapter testParameters testShortcuts testCompressedVolWriter testRawReader
```
(this is all tests that rely on Catch2 and that are available with default options)

Compiler | Before | After
----|-----|-----
GCC 10.2.0 | 9m56s | 4m15s
LLVM 10.0.0 | 6m38s | 3m30s

# Checklist

- [x] New entry in the [ChangeLog.md](https://github.com/DGtal-team/DGtal/blob/master/ChangeLog.md) added.
- [x] No warning raised in Debug ```cmake``` mode (otherwise, Travis C.I. will fail).
- [x] All continuous integration tests pass (Travis & appveyor)