@dcoeurjo 2021-02-23
before:

        make -j 20  4007,58s user 435,85s system 968% cpu 7:38,81 total

after: 

       make -j 20  3321,05s user 347,42s system 936% cpu 6:31,69 total

not exactly the same speedups but still..