@dcoeurjo 2021-02-23
Full build of DGtal + tests + examples